+++ e2e-tests/ts-describe-2/CHANGELOG.md
@@ -1,5 +1,13 @@
 # node-server
 
+## 1.0.149
+
+### Patch Changes
+
+- Updated dependencies
+  - @beff/client@0.0.152
+  - @beff/cli@0.0.152
+
 ## 1.0.148
 
 ### Patch Changes